Require y_max to always be given in nanoseconds
Changed autoplot.microbenchmark so that the unit for y_max is known at call-time. It previously took the same unit that was calculated automatically for use on the time axis, which is not known to the caller before the call to autoplot is made. Now it requires y_max to always be given in nanoseconds.